Kinda rewrite of the slot management
Basically, we don't do pushbacks anymore and the slots are always present, so we avoid a lot of issues like null checks.
We added a brand new (haha no, just a copy paste of findItem with a slightly change) method for looking empty spaces.
And the best part is, it works LMAO.